Crate fitsio

fitsio - a thin wrapper around the cfitsio C library.

File access

Pretty printing

HDU access

Creating new HDUs

Creating a new image

Creating a new table

Column descriptions

Copying HDUs to another file

Deleting a HDU

Iterating over the HDUs in a file

General calling behaviour

Header keys

Reading file data

Reading images

ndarray support

Reading tables

Reading cell values

Reading rows

Iterating over columns

Writing file data

Writing images

Resizing an image

Writing tables

Writing table data

Inserting columns

Deleting columns

Raw fits file access

Threadsafe access

This library wraps the low level cfitsio bindings: fitsio-sys and provides a more native experience for rust users.

The main interface to a fits file is FitsFile. All file manipulation and reading starts with this class.

File access

To open an existing file, use the open method.

use fitsio::FitsFile;

// let filename = ...;
let fptr = FitsFile::open(filename)?;

Alternatively a new file can be created on disk with the companion method create:

use fitsio::FitsFile;

// let filename = ...;
let fptr = FitsFile::create(filename).open()?;

The create method returns a NewFitsFile, which is an internal representation of a temporary fits file on disk, before the file is fully created.

This representation has two methods: open and with_custom_primary. The open method actually creates the file on disk, but before calling this method, the with_custom_primary method can be used to add a custom primary HDU. This is mostly useful for images. Otherwise, a default primary HDU is created. An example of not adding a custom primary HDU is shown above. Below we see an example of with_custom_primary:

use fitsio::FitsFile;
use fitsio::images::{ImageType, ImageDescription};

// let filename = ...;
let description = ImageDescription {
    data_type: ImageType::Double,
    dimensions: &[52, 103],
};

let fptr = FitsFile::create(filename)
    .with_custom_primary(&description)
    .open()?;

From this point, the current HDU can be queried and changed, or fits header cards can be read or file contents can be read.

To open a fits file in read/write mode (to allow changes to the file), the edit must be used. This opens a file which already exists on disk for editing.

use fitsio::FitsFile;

// let filename = ...;
let fptr = FitsFile::edit(filename)?;

Pretty printing

Fits files can be pretty-printed with pretty_print, or its more powerful cousin pretty_write.

use fitsio::FitsFile;

let mut fptr = FitsFile::open(filename)?;
fptr.pretty_print()?;
// or
fptr.pretty_write(&mut io::stdout())?;

In the continuing tradition of releasing fits summary programs with each fits library, this create contains a binary program fitssummary which can be installed with cargo install. This takes fits files on the command line and prints their summaries to stdout.

$ fitssummary ../testdata/full_example.fits

  file: ../testdata/full_example.fits
  mode: READONLY
  extnum hdutype      hduname    details
  0      IMAGE_HDU               dimensions: [100, 100], type: Long
  1      BINARY_TBL   TESTEXT    num_cols: 4, num_rows: 50

HDU access

HDU information belongs to the FitsHdu object. HDUs can be fetched by String/str or integer (0-indexed), with the hdu method. The HduInfo object contains information about the current HDU:

use fitsio::hdu::HduInfo;

let hdu = fptr.hdu(0)?;
// image HDU
if let HduInfo::ImageInfo { shape, .. } = hdu.info {
   println!("Image is {}-dimensional", shape.len());
   println!("Found image with shape {:?}", shape);
}

// tables
if let HduInfo::TableInfo { column_descriptions, num_rows, .. } = hdu.info {
    println!("Table contains {} rows", num_rows);
    println!("Table has {} columns", column_descriptions.len());
}

The primary HDU can always be accessed with the FitsFile::primary_hdu method.

Creating new HDUs

Creating a new image

New fits images are created with the create_image method. This method requires the extension name, and an ImageDescription object, which defines the shape and type of the desired image:

use fitsio::images::{ImageDescription, ImageType};

let image_description = ImageDescription {
    data_type: ImageType::Float,
    dimensions: &[100, 100],
};
let hdu = fptr.create_image("EXTNAME".to_string(), &image_description)?;

Unlike cfitsio, the order of the dimensions of new_size follows the C convention, i.e. row-major order.

Creating a new table

Similar to creating new images, new tables are created with the create_table method. This requires an extension name, and a slice of ColumnDescriptions:

use fitsio::tables::{ColumnDescription, ColumnDataType};

let first_description = ColumnDescription::new("A")
    .with_type(ColumnDataType::Int)
    .create()?;
let second_description = ColumnDescription::new("B")
    .with_type(ColumnDataType::Long)
    .create()?;
let descriptions = [first_description, second_description];
let hdu = fptr.create_table("EXTNAME".to_string(), &descriptions)?;

Column descriptions

Columns are described with the ColumnDescription struct. This encapsulates: the name of the column, and the data format.

The fits specification allows scalar or vector columns, and the data format is described the ColumnDataDescription struct, which in turn encapsulates the number of elements per row element (typically 1), the width of the column (for strings), and the data type, which is one of the ColumnDataType members

For the common case of a scalar column, a ColumnDataDescription object can be constructed with the scalar method:

use fitsio::tables::{ColumnDescription, ColumnDataDescription, ColumnDataType};

let desc = ColumnDataDescription::scalar(ColumnDataType::Int);
assert_eq!(desc.repeat, 1);
assert_eq!(desc.width, 1);

Vector columns can be constructed with the vector method:

use fitsio::tables::{ColumnDataDescription, ColumnDescription, ColumnDataType};

let desc = ColumnDataDescription::vector(ColumnDataType::Int, 100);
assert_eq!(desc.repeat, 100);
assert_eq!(desc.width, 1);

These impl From<...> for String such that the traditional fits column description string can be obtained:

use fitsio::tables::{ColumnDataDescription, ColumnDescription, ColumnDataType};

let desc = ColumnDataDescription::scalar(ColumnDataType::Int);
assert_eq!(String::from(desc), "1J".to_string());

Copying HDUs to another file

A HDU can be copied to another open file with the copy_to method. This requires another open FitsFile object to copy to:

hdu.copy_to(&mut src_fptr, &mut dest_fptr)?;

Deleting a HDU

The current HDU can be deleted using the delete method. Note: this method takes ownership of self, and as such the FitsHdu object cannot be used after this is called.

// let fptr = FitsFile::open(...)?;
// let hdu = fptr.hdu(0)?;
hdu.delete(&mut fptr)?;
// Cannot use hdu after this

Iterating over the HDUs in a file

The iter method allows for iteration over the HDUs of a fits file.

for hdu in fptr.iter() {
    // Do something with hdu
}

General calling behaviour

All subsequent data acess is performed through the FitsHdu object. Most methods take the currently open FitsFile as the first parameter.

Header keys

Header keys are read through the read_key function, and is generic over types that implement the ReadsKey trait:

let int_value: i64 = fptr.hdu(0)?.read_key(&mut fptr, "INTTEST")?;

// Alternatively
let int_value = fptr.hdu(0)?.read_key::<i64>(&mut fptr, "INTTEST")?;

// Or let the compiler infer the types (if possible)

HeaderValue also implements the ReadsKey trait, and allows the reading of comments:

let int_value_with_comment: HeaderValue<i64> = fptr.hdu(0)?.read_key(&mut fptr, "INTTEST")?;
let HeaderValue { value, comment } = int_value_with_comment;

Header cards can be written through the method write_key. It takes a key name and value, or a key name and value-comment tuple. See the WritesKey trait for supported data types.

fptr.hdu(0)?.write_key(&mut fptr, "foo", 1i64)?;
assert_eq!(fptr.hdu(0)?.read_key::<i64>(&mut fptr, "foo")?, 1i64);

// with comments
fptr.hdu(0)?.write_key(&mut fptr, "bar", (1i64, "bar comment"))?;

let HeaderValue { value, comment } = fptr.hdu(0)?.read_key::<HeaderValue<i64>>(&mut fptr, "bar")?;
assert_eq!(value, 1i64);
assert_eq!(comment, Some("bar comment".to_string()));

Reading file data

Methods taking ranges are exclusive of the upper range value, reflecting the nature of Rust’s range type.

Reading images

Image data can be read through either read_section which reads contiguous pixels between a start index and end index, or read_region which reads rectangular chunks from the image.

// Read the first 100 pixels
let first_row: Vec<i32> = hdu.read_section(&mut fptr, 0, 100)?;

// Read a square section of the image
let xcoord = 0..10;
let ycoord = 0..10;
let chunk: Vec<i32> = hdu.read_region(&mut fptr, &[&ycoord, &xcoord])?;

Unlike cfitsio, the order of the the section ranges follows the C convention, i.e. row-major order.

Some convenience methods are available for reading rows of the image. This is typically useful as it’s an efficient access method:

let start_row = 0;
let num_rows = 10;
let first_few_rows: Vec<f32> = hdu.read_rows(&mut fptr, start_row, num_rows)?;

// 10 rows of 100 columns
assert_eq!(first_few_rows.len(), 1000);

The whole image can also be read into memory:

let image_data: Vec<f32> = hdu.read_image(&mut fptr, )?;

// 100 rows of 100 columns
assert_eq!(image_data.len(), 10_000);

ndarray support

When fitsio is compiled with the array feature, images can be read into the ndarray::ArrayD type:

use fitsio::FitsFile;
use ndarray::ArrayD;

let mut f = FitsFile::open("../testdata/full_example.fits").unwrap();
let hdu = f.primary_hdu().unwrap();

let data: ArrayD<u32> = hdu.read_image(&mut f).unwrap();
let dim = data.dim();
assert_eq!(dim[0], 100);
assert_eq!(dim[1], 100);
assert_eq!(data[[20, 5]], 152);

For more details, see the ndarray_compat documentation (only available if compiled with array feature).

Reading tables

Columns can be read using the read_col function, which can convert data types on the fly. See the ReadsCol trait for supported data types.

let integer_data: Vec<i32> = hdu.and_then(|hdu| hdu.read_col(&mut fptr, "intcol"))?;

Reading cell values

Individual cell values can be read from FITS tables:

let result: i64 = tbl_hdu.read_cell_value(&mut f, "intcol", 4)?;
assert_eq!(result, 16);

let result: String = tbl_hdu.read_cell_value(&mut f, "strcol", 4)?;
assert_eq!(result, "value4".to_string());

Reading rows

Single rows can be read from a fits table with the row method. This requires use of the fitsio-derive crate.

use fitsio::tables::FitsRow;
use fitsio_derive::FitsRow;

#[derive(Default, FitsRow)]
struct Row {
    #[fitsio(colname = "intcol")]
    intfoo: i32,
    #[fitsio(colname = "strcol")]
    foobar: String,
}

// Pick the 4th row
let row: Row = hdu.row(&mut f, 4)?;
assert_eq!(row.intfoo, 16);
assert_eq!(row.foobar, "value4");

Iterating over columns

Iterate over the columns with columns.

for column in hdu.columns(&mut fptr) {
    // Do something with column
}

Writing file data

Methods taking ranges are exclusive of the upper range value, reflecting the nature of Rust’s range type.

Writing images

Image data is written through three methods on the HDU object: write_section, write_region, and write_image.

write_section requires a start index and end index and data to write. The data parameter needs to be a slice, meaning any contiguous memory storage method (e.g. Vec) can be passed.

let data_to_write: Vec<f64> = vec![1.0, 2.0, 3.0];
hdu.write_section(&mut fptr, 0, data_to_write.len(), &data_to_write)?;

write_region takes a slice of ranges with which the data is to be written, and the data to write.

let data_to_write: Vec<f64> = vec![1.0, 2.0, 3.0, 4.0];
let ranges = [&(0..1), &(0..1)];
hdu.write_region(&mut fptr, &ranges, &data_to_write)?;

Unlike cfitsio, the order of the ranges follows the C convention, i.e. row-major order.

write_image writes all of the data passed (if possible) into the image. If more data is passed than pixels in the image, the method returns with an error.

// Image is 3x1
assert!(hdu.write_image(&mut fptr, &[1.0, 2.0, 3.0]).is_ok());
assert!(hdu.write_image(&mut fptr, &[1.0, 2.0, 3.0, 4.0]).is_err());

Resizing an image

Images can be resized to a new shape using the resize method.

The method takes the open FitsFile, and an slice of usize values. Note: currently fitsio only supports slices with length 2, i.e. a 2D image. resize takes ownership self to force the user to fetch the HDU object again. This ensures the image changes are reflected in the hew HDU object.

use fitsio::hdu::HduInfo;

hdu.resize(&mut fptr, &[1024, 1024])?;

// Have to get the HDU again, to reflect the latest changes
let hdu = fptr.hdu(0)?;
match hdu.info {
    HduInfo::ImageInfo { shape, .. } => {
        assert_eq!(shape, [1024, 1024]);
    }
    _ => panic!("Unexpected hdu type"),
}

Unlike cfitsio, the order of the dimensions of new_size follows the C convention, i.e. row-major order.

Writing tables

Writing table data

Tablular data can either be written with write_col or write_col_range.

write_col writes an entire column’s worth of data to the file. It does not check how many rows are in the file, but extends the table if the length of data is longer than the table length.

let data_to_write: Vec<i32> = vec![10101; 5];
hdu.write_col(&mut fptr, "bar", &data_to_write)?;
let data: Vec<i32> = hdu.read_col(&mut fptr, "bar")?;
assert_eq!(data, vec![10101, 10101, 10101, 10101, 10101]);

write_col_range writes data to a range of rows in a table. The range is inclusive of both the upper and lower bounds, so 0..4 writes 5 elements.

let data_to_write: Vec<i32> = vec![10101; 10];
hdu.write_col_range(&mut fptr, "bar", &data_to_write, &(0..5))?;
let data: Vec<i32> = hdu.read_col(&mut fptr, "bar")?;
assert_eq!(data, vec![10101, 10101, 10101, 10101, 10101]);

Inserting columns

Two methods on the HDU object allow for adding new columns: append_column and insert_column. append_column adds a new column as the last column member, and is generally preferred as it does not require shifting of data within the file.

use fitsio::tables::{ColumnDescription, ColumnDataType};

let column_description = ColumnDescription::new("abcdefg")
    .with_type(ColumnDataType::Int)
    .create()?;
hdu.append_column(&mut fptr, &column_description)?;

Deleting columns

The HDU object has the method delete_column which removes a column. The column can either be accessed by integer or name

let newhdu = hdu.delete_column(&mut fptr, "bar")?;
// or
let newhdu = hdu.delete_column(&mut fptr, 0)?;

Raw fits file access

Converting a FitsFile to a raw fitsio_sys::fitsfile pointer

If this library does not support the particular use case that is needed, the raw fitsfile pointer can be accessed:

use fitsio::FitsFile;

let mut fptr = FitsFile::open(filename)?;

/* Find out the number of HDUs in the file */
let mut num_hdus = 0;
let mut status = 0;

unsafe {
    let fitsfile = fptr.as_raw();

    /* Use the unsafe fitsio-sys low level library to call a function that is possibly not
    implemented in this crate */
    fitsio_sys::ffthdu(fitsfile, &mut num_hdus, &mut status);
}
assert_eq!(num_hdus, 2);

This (unsafe) pointer can then be used with the underlying fitsio-sys library directly.

Creating a FitsFile from a raw fitsio_sys::fitsfile pointer

The inverse of the process described above can be performed. Note: calling this FitsFile constructor is unsafe => it is up to the caller to guarantee that the pointer given was

created by cfitsio (or fitsio_sys), and

it represents a valid FITS file.

Given these two things, a FitsFile can be created.

use fitsio::{sys::ffopen, FileOpenMode, FitsFile};

let filename = "../testdata/full_example.fits";
let mut fptr = std::ptr::null_mut();
let mut status = 0;
let c_filename = std::ffi::CString::new(filename).expect("filename is not a valid C-string");

unsafe {
    ffopen(
        &mut fptr as *mut *mut _,
        c_filename.as_ptr(),
        0, // readonly
        &mut status,
    );
}
assert_eq!(status, 0);

let mut f = unsafe { FitsFile::from_raw(fptr, FileOpenMode::READONLY) }.unwrap();

Threadsafe access

Access to a FitsFile is not threadsafe. Behind the scenes, fetching a FitsHdu changes internal state, and fitsio does not provide any concurrent access gauruntees. Therefore, a FitsFile does not implement Send or Sync.

In order to allow for threadsafe access, the FitsFile struct has a threadsafe method, which returns a threadsafe ThreadsafeFitsFile struct (a tuple-type wrapper around Arc<Mutex<FitsFile>>) which can be shared between threads safely.

The same concerns with Arc<Mutex<T>> data should be applied here. Additionally, the library is subject to OS level limits, such as the maximum number of open files.

Example

let f = fptr.threadsafe();

/* Spawn loads of threads... */
for i in 0..100 {
let mut f1 = f.clone();

/* Send the cloned ThreadsafeFitsFile to another thread */
let handle = thread::spawn(move || {
/* Get the underlyng fits file back */
let mut t = f1.lock().unwrap();

/* Fetch a different HDU per thread */
let hdu_num = i % 2;
let _hdu = t.hdu(hdu_num).unwrap();
});
}

Re-exports

pub use crate::headers::HeaderValue;

pub use fitsio_sys as sys;

Modules

errors

Errors and error handling

hdu

Fits HDU related code

headers

Header-related code

images

Image related code

tables

Table-related code

threadsafe_fitsfile

Thread-safe FitsFile struct

Structs

CfitsioVersion

Representation of the version of cfitsio used within bindings

FitsFile

Main entry point to the FITS file format

Enums

FileOpenMode

Enumeration of file open modes

Functions

cfitsio_version