Module fitsio::errors

Errors and error handling

This mostly concerns converting to and from the main error type defined in this crate: Error

Structs

FitsError

Error type

IndexError

Error raised when the user requests invalid indexes for data

Enums

Error

Enumeration of all error types

Functions

check_status

Function for chaining result types

Type Aliases

Result

Handy error type for use internally