Struct fitsio::hdu::FitsHdu

pub struct FitsHdu {
    pub info: HduInfo,
    pub number: usize,
}

Struct representing a FITS HDU

Fields
info: HduInfo

Information about the current HDU

number: usize

The HDU number within the fits file. Zero indexed.

Implementations

impl FitsHdu

pub fn name(&self, fits_file: &mut FitsFile) -> Result<String>

Read the HDU name

pub fn read_key<T: ReadsKey>( &self, fits_file: &mut FitsFile, name: &str, ) -> Result<T>

Read header key

Example

let int_value: i64 = hdu.read_key(&mut fptr, "INTTEST")?;
    

pub fn write_key<T: WritesKey>( &self, fits_file: &mut FitsFile, name: &str, value: T, ) -> Result<()>

Write a fits key to the current header

Example

fptr.primary_hdu()?.write_key(&mut fptr, "foo", 1i64)?;
assert_eq!(fptr.hdu(0)?.read_key::<i64>(&mut fptr, "foo")?, 1i64);

pub fn read_section<T: ReadImage>( &self, fits_file: &mut FitsFile, start: usize, end: usize, ) -> Result<T>

Read pixels from an image between a start index and end index

The range is exclusive of the upper value

Example

// Read the first 100 pixels
let first_row: Vec<i32> = hdu.read_section(&mut fptr, 0, 100)?;

pub fn read_rows<T: ReadImage>( &self, fits_file: &mut FitsFile, start_row: usize, num_rows: usize, ) -> Result<T>

Example

let start_row = 0;
let num_rows = 10;
let first_few_rows: Vec<f32> = hdu.read_rows(&mut fptr, start_row, num_rows)?;

// 10 rows of 100 columns
assert_eq!(first_few_rows.len(), 1000);

pub fn read_row<T: ReadImage>( &self, fits_file: &mut FitsFile, row: usize, ) -> Result<T>

Read a single row from a fits image

Example

let chosen_row = 5;
let row: Vec<f32> = hdu.read_row(&mut fptr, chosen_row)?;

// Should have 100 pixel values
assert_eq!(row.len(), 100);

pub fn read_region<T: ReadImage>( &self, fits_file: &mut FitsFile, ranges: &[&Range<usize>], ) -> Result<T>

Read a square region from the chip.

Lower left indicates the starting point of the square, and the upper right defines the pixel beyond the end. The range of pixels included is inclusive of the lower end, and exclusive of the upper end.

Example

// Read a square section of the image
let xcoord = 0..10;
let ycoord = 0..10;
let chunk: Vec<i32> = hdu.read_region(&mut fptr, &[&ycoord, &xcoord])?;

pub fn read_image<T: ReadImage>(&self, fits_file: &mut FitsFile) -> Result<T>

Read a whole image into a new Vec

This reads an entire image into a one-dimensional vector

Example

let image_data: Vec<f32> = hdu.read_image(&mut fptr)?;

// 100 rows of 100 columns
assert_eq!(image_data.len(), 10_000);

pub fn write_section<T: WriteImage>( &self, fits_file: &mut FitsFile, start: usize, end: usize, data: &[T], ) -> Result<()>

Write raw pixel values to a FITS image

If the length of the dataset exceeds the number of columns, the data wraps around to the next row.

The range is exclusive of the upper value.

Example

let data_to_write: Vec<f64> = vec![1.0, 2.0, 3.0];
hdu.write_section(&mut fptr, 0, data_to_write.len(), &data_to_write)?;

pub fn write_region<T: WriteImage>( &self, fits_file: &mut FitsFile, ranges: &[&Range<usize>], data: &[T], ) -> Result<()>

Write a rectangular region to the fits image

The ranges must have length of 2, and they represent the limits of each axis. The limits are inclusive of the lower bounds, and exclusive of the and upper bounds.

For example, writing with ranges 0..10 and 0..10 wries an 10x10 sized image.

Example

let data_to_write: Vec<f64> = vec![1.0, 2.0, 3.0, 4.0];
let ranges = [&(0..1), &(0..1)];
hdu.write_region(&mut fptr, &ranges, &data_to_write)?;

pub fn write_image<T: WriteImage>( &self, fits_file: &mut FitsFile, data: &[T], ) -> Result<()>

Write an entire image to the HDU passed in

Firstly a check is performed, making sure that the amount of data will fit in the image. After this, all of the data is written to the image.

Example

// Image is 3x1
assert!(hdu.write_image(&mut fptr, &[1.0, 2.0, 3.0]).is_ok());
assert!(hdu.write_image(&mut fptr, &[1.0, 2.0, 3.0, 4.0]).is_err());

pub fn resize( self, fits_file: &mut FitsFile, new_size: &[usize], ) -> Result<FitsHdu>

Resize a HDU image

The new_size parameter defines the new size of the image. Unlike cfitsio, the order of the dimensions of new_size follows the C convention, i.e. row-major order.

Example

use fitsio::hdu::HduInfo;

hdu.resize(&mut fptr, &[1024, 1024])?;
// Have to get the HDU again, to reflect the latest changes
let hdu = fptr.hdu(0)?;
match hdu.info {
    HduInfo::ImageInfo { shape, .. } => {
        assert_eq!(shape, [1024, 1024]);
    }
    _ => panic!("Unexpected hdu type"),
}

pub fn copy_to( &self, src_fits_file: &mut FitsFile, dest_fits_file: &mut FitsFile, ) -> Result<()>

Copy an HDU to another open fits file

Example

hdu.copy_to(&mut src_fptr, &mut dest_fptr)?;

pub fn insert_column( self, fits_file: &mut FitsFile, position: usize, description: &ConcreteColumnDescription, ) -> Result<FitsHdu>

Insert a column into a fits table

The column location is 0-indexed. It is inserted at that position, and the following columns are shifted back.

Example

use fitsio::tables::{ColumnDescription, ColumnDataType};

let column_description = ColumnDescription::new("abcdefg")
    .with_type(ColumnDataType::Int)
    .create()?;
hdu.insert_column(&mut fptr, 1, &column_description)?;

pub fn append_column( self, fits_file: &mut FitsFile, description: &ConcreteColumnDescription, ) -> Result<FitsHdu>

Add a new column to the end of the table

Example

use fitsio::tables::{ColumnDescription, ColumnDataType};

let column_description = ColumnDescription::new("abcdefg")
    .with_type(ColumnDataType::Int)
    .create()?;
hdu.append_column(&mut fptr, &column_description)?;

pub fn delete_column<T: DescribesColumnLocation>( self, fits_file: &mut FitsFile, col_identifier: T, ) -> Result<FitsHdu>

Remove a column from the fits file

The column can be identified by id or name.

Example

let newhdu = hdu.delete_column(&mut fptr, "bar")?;
// or
let newhdu = hdu.delete_column(&mut fptr, 0)?;

pub fn read_col<T: ReadsCol>( &self, fits_file: &mut FitsFile, name: &str, ) -> Result<Vec<T>>

Read a subset of a fits column

The range is exclusive of the upper value

Example

let data_to_write: Vec<i32> = vec![10101; 10];
hdu.write_col_range(&mut fptr, "bar", &data_to_write, &(0..5))?;
let data: Vec<i32> = hdu.read_col(&mut fptr, "bar")?;
assert_eq!(data, vec![10101, 10101, 10101, 10101, 10101]);

pub fn read_col_range<T: ReadsCol>( &self, fits_file: &mut FitsFile, name: &str, range: &Range<usize>, ) -> Result<Vec<T>>

Read a subset of a fits column

The range is exclusive of the upper value

Example

let data: Vec<i32> = hdu.read_col_range(&mut fptr, "bar", &(0..5))?;
assert_eq!(data, vec![10101, 10101, 10101, 10101, 10101]);

pub fn write_col_range<T: WritesCol, N: Into<String>>( &self, fits_file: &mut FitsFile, name: N, col_data: &[T], rows: &Range<usize>, ) -> Result<FitsHdu>

Write data to part of a column

The range is exclusive of the upper value

Example

let data_to_write: Vec<i32> = vec![10101; 10];
hdu.write_col_range(&mut fptr, "bar", &data_to_write, &(0..5))?;

pub fn write_col<T: WritesCol, N: Into<String>>( &self, fits_file: &mut FitsFile, name: N, col_data: &[T], ) -> Result<FitsHdu>

Write data to an entire column

This default implementation does not check the length of the column first, but if the length of the data array is longer than the length of the table, the table will be extended with extra rows. This is as per the fitsio definition.

Example

let data_to_write: Vec<i32> = vec![10101; 5];
hdu.write_col(&mut fptr, "bar", &data_to_write)?;

pub fn columns<'a>(&self, fits_file: &'a mut FitsFile) -> ColumnIterator<'a>

Iterate over the columns in a fits file

Example

for column in hdu.columns(&mut fptr) {
    // Do something with column
}

pub fn delete(self, fits_file: &mut FitsFile) -> Result<()>

Delete the current HDU from the fits file.

Note this method takes self by value, and as such the hdu cannot be used after this method is called.

Example

// let fptr = FitsFile::open(...)?;
// let hdu = fptr.hdu(0)?;
hdu.delete(&mut fptr)?;
// Cannot use hdu after this

pub fn read_cell_value<T>( &self, fits_file: &mut FitsFile, name: &str, idx: usize, ) -> Result<T>
where T: ReadsCol,

Read a single value from a fits table

This will be inefficient if lots of individual values are wanted.

Example

let result: i64 = tbl_hdu.read_cell_value(&mut f, "intcol", 4)?;
assert_eq!(result, 16);

let result: String = tbl_hdu.read_cell_value(&mut f, "strcol", 4)?;
assert_eq!(result, "value4".to_string());

pub fn row<F>(&self, fits_file: &mut FitsFile, idx: usize) -> Result<F>
where F: FitsRow,

Extract a single row from the file

This method uses returns a FitsRow, which is provided by the user, using a derive implementation from the fitsio-derive crate.

Example

use fitsio::tables::FitsRow;
use fitsio_derive::FitsRow;

#[derive(Default, FitsRow)]
struct Row {
    #[fitsio(colname = "intcol")]
    intfoo: i32,
    #[fitsio(colname = "strcol")]
    foobar: String,
}

// Pick the 4th row
let row: Row = hdu.row(&mut f, 4)?;
assert_eq!(row.intfoo, 16);
assert_eq!(row.foobar, "value4");