Struct fitsio::FitsFile

pub struct FitsFile {
    pub filename: String,
    // some fields omitted
}

Wrapper around a C fitsfile pointer.

This handles opening, creating and automatically closing (through the Drop trait) the file.

All subsequent file access is controled through this object.

Fields
filename: String
Methods

impl FitsFile

fn open(filename: &str) -> Result<Self>

Open a fits file for reading

filename - Filename to pass to cfitsio. Can conform to the Extended Filename Syntax.

Examples

    match FitsFile::open("../testdata/full_example.fits") {
        Ok(f) => { },
        Err(e) => panic!("{:?}", e),
    }

fn create(path: &str) -> Result<Self>

Create a new fits file

An empty primary header is added so when the drop method is called, the file is valid.

Examples

match FitsFile::create(filename) {
    Ok(f) => println!("Fits file created ok"),
    Err(e) => panic!("Error: {:?}", e),
}

fn current_hdu_number(&self) -> usize

Returns the current HDU number, 0-indexed

The FITS standard is 1-indexed (thanks Fortran), so this function converts the indexing to be 0-based so the primary HDU is at index 0.

Examples

let f = FitsFile::open("../testdata/full_example.fits").unwrap();
assert_eq!(f.current_hdu_number(), 0);

fn change_hdu<T: DescribesHdu>(&self, hdu_description: T)

Change the current HDU

Examples

assert_eq!(f.current_hdu_number(), 0);
f.change_hdu(1);
assert_eq!(f.current_hdu_number(), 1);
f.change_hdu(0);
assert_eq!(f.current_hdu_number(), 0);
f.change_hdu("TESTEXT");
f.change_hdu(1);

fn get_hdu_type(&self) -> FitsHduType

Get which type of HDU the current HDU is

Results in one of the FitsHduType options.

Examples

let mut f = FitsFile::open("../testdata/full_example.fits").unwrap();
// Primary HDUs are always image hdus
assert_eq!(f.get_hdu_type(), FitsHduType::ImageHDU);

fn get_hdu(&self, index: usize) -> FitsHDU

Return a FitsHDU object for the specified HDU

Examples

let primary_hdu = f.get_hdu(0);
assert_eq!(primary_hdu.hdu_type, FitsHduType::ImageHDU);

fn read_key<T: ReadsKey>(&self, name: &str) -> Result<T>

Function to read a header key from the current HDU

The function is generic on the trait ReadsKey which is implemented for multiple data types (see reads_key_impl).