Struct fitsio::FitsFile

pub struct FitsFile { /* private fields */ }

Main entry point to the FITS file format

Implementations

impl FitsFile

pub fn open<T: AsRef<Path>>(filename: T) -> Result<Self>

Open a fits file from disk

Example

use fitsio::FitsFile;

// let filename = ...;
let fptr = FitsFile::open(filename)?;

pub fn edit<T: AsRef<Path>>(filename: T) -> Result<Self>

Open a fits file in read/write mode

Example

use fitsio::FitsFile;

// let filename = ...;
let fptr = FitsFile::edit(filename)?;

pub fn create<'a, T: AsRef<Path>>(path: T) -> NewFitsFile<'a, T>

Create a new fits file on disk

The create method returns a NewFitsFile, which is an internal representation of a temporary fits file on disk, before the file is fully created.

This representation has two methods: open and with_custom_primary. The open method actually creates the file on disk, but before calling this method, the with_custom_primary method can be used to add a custom primary HDU. This is mostly useful for images. Otherwise, a default primary HDU is created. An example of not adding a custom primary HDU is shown above. Below we see an example of with_custom_primary:

Example

use fitsio::FitsFile;
use fitsio::images::{ImageDescription, ImageType};

// let filename = ...;
let description = ImageDescription {
    data_type: ImageType::Double,
    dimensions: &[52, 103],
};

let fptr = FitsFile::create(filename)
    .with_custom_primary(&description)
    .open()?;

pub fn hdu<T: DescribesHdu>(&mut self, hdu_description: T) -> Result<FitsHdu>

Return a new HDU object

HDU information belongs to the FitsHdu object. HDUs can be fetched by String/str or integer (0-indexed). The HduInfo object contains information about the current HDU:

Example

use fitsio::hdu::HduInfo;
let hdu = fptr.hdu(0)?;
// image HDU
if let HduInfo::ImageInfo { shape, .. } = hdu.info {
   println!("Image is {}-dimensional", shape.len());
   println!("Found image with shape {:?}", shape);
}

// tables
if let HduInfo::TableInfo { column_descriptions, num_rows, .. } = hdu.info {
    println!("Table contains {} rows", num_rows);
    println!("Table has {} columns", column_descriptions.len());
}

pub fn primary_hdu(&mut self) -> Result<FitsHdu>

Return the primary hdu (HDU 0)

Example

let hdu = fptr.hdu(0)?;
let phdu = fptr.primary_hdu()?;
assert_eq!(hdu, phdu);

pub fn file_path(&self) -> &Path

Return the file path of the file

pub fn create_table<T>( &mut self, extname: T, table_description: &[ConcreteColumnDescription], ) -> Result<FitsHdu>
where T: Into<String>,

Create a new fits table

Create a new fits table, with columns as detailed in the ColumnDescription object.

Example

use fitsio::tables::{ColumnDataType, ColumnDescription};

let first_description = ColumnDescription::new("A")
    .with_type(ColumnDataType::Int)
    .create()?;
let second_description = ColumnDescription::new("B")
    .with_type(ColumnDataType::Long)
    .create()?;
let descriptions = &[first_description, second_description];
let hdu = fptr.create_table("EXTNAME".to_string(), descriptions)?;

pub fn create_image<T>( &mut self, extname: T, image_description: &ImageDescription<'_>, ) -> Result<FitsHdu>
where T: Into<String>,

Create a new fits image, and return the FitsHdu object.

This method takes an ImageDescription struct which defines the desired layout of the image HDU.

Example

use fitsio::images::{ImageDescription, ImageType};

let image_description = ImageDescription {
    data_type: ImageType::Float,
    dimensions: &[100, 100],
};
let hdu = fptr.create_image("EXTNAME".to_string(), &image_description)?;

pub fn iter(&mut self) -> FitsHduIterator<'_>

Iterate over the HDUs in the file

Example

for hdu in fptr.iter() {
    // Do something with hdu
}

pub fn pretty_print(&mut self) -> Result<()>

Pretty-print file to stdout

Fits files can be pretty-printed with pretty_print, or its more powerful cousin pretty_write.

Example

use fitsio::FitsFile;

let mut fptr = FitsFile::open(filename)?;
fptr.pretty_print()?;
// or
fptr.pretty_write(&mut io::stdout())?;

pub fn pretty_write<W>(&mut self, w: &mut W) -> Result<()>
where W: Write,

Pretty-print the fits file structure to any Write implementor

Fits files can be pretty-printed with pretty_print, or its more powerful cousin pretty_write.

Example

use fitsio::FitsFile;

let mut fptr = FitsFile::open(filename)?;
fptr.pretty_print()?;
// or
fptr.pretty_write(&mut io::stdout())?;

pub unsafe fn as_raw(&mut self) -> *mut fitsfile

Return a pointer to the underlying C fitsfile object representing the current file.

Any changes to the underlying fits file will not be updated in existing FitsHdu objects, so these must be recreated.

Safety

This is marked as unsafe as it is definitely something that is not required by most users, and hence the unsafe-ness marks it as an advanced feature. I have also not considered possible concurrency or data race issues as yet.

Example

use fitsio::{sys, FitsFile};

let mut fptr = FitsFile::open(filename)?;

/* Find out the number of HDUs in the file */
let mut num_hdus = 0;
let mut status = 0;

unsafe {
    let fitsfile = fptr.as_raw();

    /* Use the unsafe fitsio-sys low level library to call a function that is possibly not
    implemented in this crate */
    fitsio_sys::ffthdu(fitsfile, &mut num_hdus, &mut status);
}
assert_eq!(num_hdus, 2);

pub unsafe fn from_raw( fptr: *mut fitsfile, mode: FileOpenMode, ) -> Result<FitsFile>

Load a FitsFile from a fitsio_sys::fitsfile pointer.

Safety

This constructor is inherently unsafe - the Rust compiler cannot verify the validity of the pointer supplied. It is therefore the responsibility of the caller to prove that the pointer is a valid fitsfile by calling an unsafe function.

it is up to the caller to guarantee that the pointer given was

created by cfitsio (or fitsio_sys), and

it represents a valid FITS file.

Given these two things, a FitsFile can be created.

Example

use fitsio::{sys::ffopen, FileOpenMode, FitsFile};

let filename = "../testdata/full_example.fits";
let mut fptr = std::ptr::null_mut();
let mut status = 0;
let c_filename = std::ffi::CString::new(filename).expect("filename is not a valid C-string");

unsafe {
    ffopen(
        &mut fptr as *mut *mut _,
        c_filename.as_ptr(),
        0, // readonly
        &mut status,
    );
}
assert_eq!(status, 0);

let mut f = unsafe { FitsFile::from_raw(fptr, FileOpenMode::READONLY) }.unwrap();

impl FitsFile

pub fn threadsafe(self) -> ThreadsafeFitsFile

Create a threadsafe ThreadsafeFitsFile copy of the current FitsFile.